refactor(web): remove backward-compat aliases from eventListStore and move to features/streams
