+++ apps/web/src/features/menu/LiveFilterButton.vue
@@ -5,7 +5,7 @@
 <script setup lang="ts">
 import { computed } from "vue";
 import { useSearchStore } from "../../shared/stores";
-import { useEventListStore } from "../../store/eventListStore";
+import { useEventListStore } from "../streams";
 
 const searchStore = useSearchStore();
 const eventListStore = useEventListStore();
