+++ apps/web/src/features/streams/index.ts
@@ -1,2 +1,3 @@
+export { useEventListStore } from "./eventListStore";
 export { default as LiverEventDateSectionList } from "./LiverEventDateSectionList.vue";
 export { default as LiverEventDetailList } from "./detail/LiverEventDetailList.vue";
